requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
et_xmlfile==2.0.0
gunicorn==25.1.0
joblib==1.5.3
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
whitenoise==6.12.0
